Add Swift C++ interop annotations for OpenUSD 25.11 compatibility
- Add SWIFT_SHARED_REFERENCE to UsdStage and SdfLayer with retain/release functions
- Add SWIFT_IMMORTAL_REFERENCE to Hgi, HgiGL, HgiMetal, ArResolver, ArDefaultResolver
- Add SWIFT_IMMORTAL_REFERENCE to UsdValidationRegistry, UsdImagingGLEngine
- Convert UsdListPosition and UsdLoadPolicy from enum to enum class
- Update Swift wrapper code for new enum naming conventions
- Update XformableMacros precision enum names

Following wabiverse SwiftUSD convention for Swift C++ interoperability.

+++ Sources/Gf/math.cpp
@@ -6,14 +6,16 @@
 //
 
 #include "pxr/pxrns.h"
-#include "Gf/math.h"
+#include "Gf/gfMath.h"
+
+#include <cmath>
 
 PXR_NAMESPACE_OPEN_SCOPE
 
 double
 GfMod(double a, double b)
 {
-    double c = fmod(a, b);
+    double c = std::fmod(a, b);
     if (a < 0)
         return c ? (b + c) : 0;
     else
@@ -23,7 +25,7 @@ GfMod(double a, double b)
 float
 GfMod(float a, float b)
 {
-    double c = fmodf(a, b);
+    float c = std::fmod(a, b);
     if (a < 0)
         return c ? (b + c) : 0;
     else
